+++ src/pages/PersonalSchedulePage/PersonalSchedulePage.styles.js
@@ -1,0 +1,10 @@
+import styled from "styled-components";
+
+// 나중에 컴포넌트 추가할 수도 있으므로 export const로
+export const LayoutMain = styled.main`
+	display: flex;
+	justify-content: center;
+	gap: 13px;
+	padding: 50px 60px 0;
+	font-family: "Inter", sans-serif;
+`;
